deal_scout/requirements.txt
pandas>=2.2.0
pytest>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.1
streamlit>=1.44.0
folium>=0.16.0
streamlit-folium>=0.18.0
reportlab>=4.1.0
APScheduler>=3.10.4
gspread>=6.0.0
google-auth>=2.0.0
